Move Loader and PageLayout components
