+++ identity/webapp/views/components/MyAccount/DeleteAccount.tsx
@@ -10,11 +10,11 @@ import {
   RequestDeleteError,
   useRequestDelete,
 } from '@weco/identity/hooks/useRequestDelete';
+import Loading from '@weco/identity/views/components/Loading';
 import { PasswordInput } from '@weco/identity/views/components/PasswordInput';
 import { FieldMargin } from '@weco/identity/views/components/styled/forms';
 
 import { ChangeDetailsModalContentProps } from './ChangeDetailsModal';
-import Loading from './Loading';
 import {
   ButtonAlign,
   ModalContainer,
