+++ identity/webapp/views/pages/success.tsx
@@ -8,11 +8,11 @@ import {
 import Space from '@weco/common/views/components/styled/Space';
 import { usePageTitle } from '@weco/identity/hooks/usePageTitle';
 import { ApplicationReceived } from '@weco/identity/utils/copy';
-import PageWrapper from '@weco/identity/views/components/PageWrapper';
 import {
   Container,
   Wrapper,
 } from '@weco/identity/views/components/styled/layouts';
+import IdentityPageLayout from '@weco/identity/views/layouts/IdentityPageLayout';
 
 export type Props = {
   email: string;
@@ -22,7 +22,7 @@ const SuccessPage: NextPage<Props> = ({ email }) => {
   usePageTitle('Application received');
 
   return (
-    <PageWrapper title="Registration">
+    <IdentityPageLayout title="Registration">
       <ContaineredLayout gridSizes={gridSize10()}>
         <Space $v={{ size: 'xl', properties: ['margin-top'] }}>
           <Container>
@@ -36,7 +36,7 @@ const SuccessPage: NextPage<Props> = ({ email }) => {
           </Container>
         </Space>
       </ContaineredLayout>
-    </PageWrapper>
+    </IdentityPageLayout>
   );
 };
